sdk: migrating over to the renamed security sdk
